requirements.txt
asgiref==3.8.1
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
sqlparse==0.4.4
typing_extensions==4.10.0
requests==2.31.0
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2023.3
toml==0.10.2
tomli==2.0.1
tzdata==2023.3
pillow==10.3.0
django-extra-fields==0.3
